requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
email_validator==2.1.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
SQLAlchemy==2.0.24
Werkzeug==2.3.7
python-dotenv==1.0.0

# These are optional depending on your production database
# psycopg2-binary==2.9.9  # For PostgreSQL
# mysqlclient==2.1.1      # For MySQL 